Extract to a `colours` module rather than in palette16
