+++ agb/src/display/mod.rs
@@ -10,10 +10,12 @@ use self::{
     window::Windows,
 };
 
-pub use palette16::{Palette16, Rgb, Rgb15, include_palette};
+pub use colours::{Rgb, Rgb15};
+pub use palette16::{Palette16, include_palette};
 
 /// Graphics mode 3. Bitmap mode that provides a 16-bit colour framebuffer.
 pub(crate) mod bitmap3;
+mod colours;
 /// Test logo of agb.
 pub mod example_logo;
 pub mod object;
